Clean up add location
